estrutura fix
