+++ Gestao Para Hortifruti Vs0.03/include/menu.h
@@ -1,0 +1,6 @@
+#ifndef MENU_H
+#define MENU_H
+
+void EntradaMercadoria();
+
+#endif
